Adding RGB segmentation tool
Relies on tree_routines_X3 for execution. Takes RAW RGB files from Tree Surveyor, and a list of tips. It will extract the tips, align them vertically. It will write them to disk as raw 8bit files, and a text file to set up conversion of the files to MINC format to be used with the MiCE Toronto image registration tools.